renamed users migration/model file
